requirements.txt
Django==1.5
django-plugins==0.2.2
South
coveralls
-e git://github.com/mysociety/popit-django.git#egg=popit-django
mock
django-tastypie
django-nose
email_reply_parser
flufl.bounce
django-markdown-deux
requests
django-extensions
django-subdomains
django-haystack
pysolr